@@ -13,16 +13,17 @@
   "author": "Anton Borkovskij",
   "license": "ISC",
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.0",
+    "@reduxjs/toolkit": "^1.8.1",
+    "@types/react-redux": "^7.1.24",
     "@types/uuid": "^8.3.4",
     "css-loader": "^6.7.1",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
-    "react-redux": "^7.2.6",
+    "react-redux": "^8.0.0",
     "react-router-dom": "^6.2.2",
     "react-toastify": "^8.2.0",
     "redux-saga": "^1.1.3",
@@ -38,8 +39,6 @@
     "@babel/preset-typescript": "^7.16.7",
     "@babel/runtime": "^7.17.2",
     "@testing-library/react": "^12.1.4",
-    "@types/react": "^17.0.40",
-    "@types/react-dom": "^17.0.13",
     "@types/styled-components": "^5.1.24",
     "@types/webpack": "^5.28.0",
     "@types/webpack-dev-server": "^4.7.2",
